refactor: 统一浮点类型为 f32，移除 f64 过度设计
将 LeakyReLU API 及相关测试中的 f64 参数类型统一为 f32，
符合项目 CPU 高效运行的 FP32 统一设计原则。

+++ src/nn/graph/inner/node_builders.rs
@@ -369,7 +369,7 @@ impl GraphInner {
     pub fn new_leaky_relu_node(
         &mut self,
         parent_id: NodeId,
-        negative_slope: f64,
+        negative_slope: f32,
         name: Option<&str>,
     ) -> Result<NodeId, GraphError> {
         let handle = NodeHandle::new_leaky_relu(&self.get_nodes(&[parent_id])?, negative_slope)?;
